requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
cattrs==1.10.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
h11==0.13.0
idna==3.3
mcpi==1.2.1
moodlepy==0.23.7
pydantic==1.9.1
requests==2.27.1
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
